Complete A-Assertions
Classes do not have any checks for their invariants
or assumptions made.

Not checking for these invariants or assumptions
will cause the program to crash or give incorrect
results if these invariants or assumptions are not
met.

Add assertions to check for invariants in the code.

Assertions is preferred as it gets the job done and it
can be turned on or off.